refactor: 💡 export-file general API to export-asset specific AP
I

✅ Closes: #3181
