+++ src/constants/api_connection.ts
@@ -146,7 +146,7 @@ export enum APIName {
   ASSET_SUGGESTED_NUMBER_GET_BY_TYPE = 'ASSET_SUGGESTED_NUMBER_GET_BY_TYPE',
   TRIAL_BALANCE_LIST = 'TRIAL_BALANCE_LIST',
   IP_LIST = 'IP_LIST',
-  FILE_EXPORT = 'FILE_EXPORT',
+  ASSET_LIST_EXPORT = 'ASSET_LIST_EXPORT',
 }
 
 export enum APIPath {
@@ -271,7 +271,7 @@ export enum APIPath {
   ASSET_SUGGESTED_NUMBER_GET_BY_TYPE = `${apiPrefixV2}/company/:companyId/asset/suggested_number`,
   TRIAL_BALANCE_LIST = `${apiPrefixV2}/company/:companyId/trial_balance`,
   IP_LIST = `${apiPrefixV2}/user/:userId/ip`,
-  FILE_EXPORT = `${apiPrefixV2}/company/:companyId/export`,
+  ASSET_LIST_EXPORT = `${apiPrefixV2}/company/:companyId/asset/export`,
 }
 const createConfig = ({
   name,
@@ -889,9 +889,9 @@ export const APIConfig: Record<IAPIName, IAPIConfig> = {
     method: HttpMethod.GET,
     path: APIPath.IP_LIST,
   }),
-  [APIName.FILE_EXPORT]: createConfig({
-    name: APIName.FILE_EXPORT,
+  [APIName.ASSET_LIST_EXPORT]: createConfig({
+    name: APIName.ASSET_LIST_EXPORT,
     method: HttpMethod.GET,
-    path: APIPath.FILE_EXPORT,
+    path: APIPath.ASSET_LIST_EXPORT,
   }),
 };
