+++ src/constants/export_asset.ts
@@ -4,10 +4,6 @@ export enum ExportFileType {
   CSV = 'csv',
 }
 
-export enum ExportType {
-  ASSETS = 'assets',
-}
-
 export enum AssetSortBy {
   ACQUISITION_DATE = 'acquisitionDate',
   PURCHASE_PRICE = 'purchasePrice',
