+++ src/interfaces/api_connection.ts
@@ -122,7 +122,7 @@ export type IAPIName =
   | 'ASSET_SUGGESTED_NUMBER_GET_BY_TYPE'
   | 'TRIAL_BALANCE_LIST'
   | 'IP_LIST'
-  | 'FILE_EXPORT';
+  | 'ASSET_LIST_EXPORT';
 
 export type IHttpMethod = 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH' | 'HEAD';
