+++ src/interfaces/export_asset.ts
@@ -1,9 +1,8 @@
-import { AssetSortBy, ExportFileType, ExportType } from '@/constants/export_file';
+import { AssetSortBy, ExportFileType } from '@/constants/export_asset';
 import { SortOrder } from '@/constants/sort';
 
 // TODO: (20241108 - Shirley) 改成 zod input / output schema
 export interface IBaseExportRequestBody {
-  exportType: string;
   fileType: ExportFileType;
 }
 
@@ -15,7 +14,6 @@ export interface IAssetSort {
 }
 
 export interface IAssetExportRequestBody extends IBaseExportRequestBody {
-  exportType: ExportType.ASSETS;
   filters?: {
     type?: string;
     status?: string;
